feat: Remove legacy HTTP client
Legacy HTTP client is removed along with `DYNATRACE_HTTP_LEGACY` and `DYNATRACE_HTTP_RESPONSE`environment variables.
