+++ dynatrace/rest/api_token_client.go
@@ -18,7 +18,6 @@
 package rest
 
 import (
-	"github.com/dynatrace-oss/terraform-provider-dynatrace/provider/envutils"
 	"context"
 	"errors"
 	"net/http"
@@ -83,13 +82,6 @@ func (me *api_token_request) Finish(vs ...any) error {
 	credentials := me.client.Credentials()
 	if !credentials.ContainsAPIToken() {
 		return NoAPITokenError
-	}
-	if envutils.DynatraceHTTPLegacy.Get() {
-		legacyRequest := legacy_request(*me)
-		if credentials.URL == TestCaseEnvURL {
-			return errors.New("legacy")
-		}
-		return legacyRequest.Finish(vs...)
 	}
 	classicRequest := classic_request(*me)
 	if credentials.URL == TestCaseEnvURL {
